Fix a couple bugs in the Scrolling Rooms Transitions script
The camera 2d was found by name. This won't work if the camera has been renamed. Instead, pull the active camera from the player viewport's world. An unused variable was also removed.

Secondly, fix the transition timing. The screen_delta variable can be negative depending on the signs of the offsets. If this happens, either the separate x or y transition will end up with a negative time, and thus the tween skips that part of the animation. Adding an absf fixes this.